requirements-dev.txt
-i https://pypi.org/simple
exceptiongroup==1.2.0; python_version < '3.11'
flake8==7.0.0; python_full_version >= '3.8.1'
iniconfig==2.0.0; python_version >= '3.7'
mccabe==0.7.0; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
pycodestyle==2.11.1; python_version >= '3.8'
pyflakes==3.2.0; python_version >= '3.8'
pytest==7.4.4; python_version >= '3.7'
tomli==2.0.1; python_version < '3.11'
blinker==1.7.0; python_version >= '3.8'
cffi==1.16.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
cryptography==41.0.7; python_version >= '3.7'
flask==3.0.0; python_version >= '3.8'
importlib-metadata==7.0.1; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
pycparser==2.21
pyopenssl==23.2.0; python_version >= '3.6'
werkzeug==3.0.1; python_version >= '3.8'
zipp==3.17.0; python_version >= '3.8'
